+++ Service/Sources/Data/RemoteDataSource/RestAPI/DataMapping/Rank/MySchoolRankDTO.swift
@@ -2,8 +2,27 @@ import Foundation
 
 struct MySchoolRankDTO: Decodable {
     private enum CodingKeys: String, CodingKey {
-        case mySchoolRank = "my_school_rank"
+        case id = "school_id"
+        case name
+        case logoImageUrlString = "logo_image_url"
+        case grade
+        case classNum = "class_num"
     }
-    let mySchoolRank: MySchoolDTO
+    let id: Int
+    let name: String
+    let logoImageUrlString: String
+    let grade: Int?
+    let classNum: Int?
 }
 
+extension MySchoolRankDTO {
+    func toDomain() -> MySchool {
+        return .init(
+            id: id,
+            name: name,
+            logoImageUrl: URL(string: logoImageUrlString)!,
+            grade: grade ?? 0,
+            classNum: classNum ?? 0
+        )
+    }
+}
